Initial proc macro layout
