+++ ex_em_ell/Cargo.toml
@@ -7,8 +7,13 @@ description = "Provide macros for serializing and deserializing XML"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
+ex_em_ell_derive = { path = "../ex_em_ell_derive", version = "0.1.0", optional = true }
 thiserror = "1.0.44"
 xml-rs = "0.8.15"
+
+[features]
+default = ["derive"]
+derive = ["ex_em_ell_derive"]
 
 [package.metadata.cyclonedx]
 format = "json"
