+++ ex_em_ell/src/lib.rs
@@ -1,44 +1,5 @@
-use std::io::{Read, Write};
-use xml::{
-    attribute::OwnedAttribute, name::OwnedName, namespace::Namespace, EventReader, EventWriter,
-};
-
 pub mod errors;
+pub mod traits;
 
-use errors::{XmlReadError, XmlWriteError};
-
-pub trait ToXmlDocument {
-    fn to_xml_document<W: Write>(
-        self: &Self,
-        writer: &mut EventWriter<W>,
-    ) -> Result<(), XmlWriteError>;
-}
-
-pub trait ToXmlElement {
-    fn to_xml_element<W: Write>(
-        self: &Self,
-        writer: &mut EventWriter<W>,
-        tag: &str,
-    ) -> Result<(), XmlWriteError>;
-
-    fn will_write(self: &Self) -> bool {
-        true
-    }
-}
-
-pub trait FromXmlDocument {
-    fn from_xml_document<R: Read>(reader: &mut EventReader<R>) -> Result<Self, XmlReadError>
-    where
-        Self: Sized;
-}
-
-pub trait FromXmlElement {
-    fn from_xml_element<R: Read>(
-        reader: &mut EventReader<R>,
-        element_name: &OwnedName,
-        element_attributes: &[OwnedAttribute],
-        element_namespace: &Namespace,
-    ) -> Result<Self, XmlReadError>
-    where
-        Self: Sized;
-}
+pub use ex_em_ell_derive::{FromXmlDocument, FromXmlElement, ToXmlDocument, ToXmlElement};
+pub use traits::{FromXmlDocument, FromXmlElement, ToXmlDocument, ToXmlElement};
