+++ flake.nix
@@ -42,7 +42,17 @@
           version = "0.1.0";
         };
 
-        cargoArtifacts = craneLib.buildDepsOnly commonArgs;
+        cargoArtifacts = craneLib.buildDepsOnly (commonArgs // {
+          dummySrc = craneLib.mkDummySrc {
+            inherit src;
+
+            # ex_em_ell_derive is a proc macro crate, so it cannot have non-proc macro functions
+            extraDummyScript = ''
+              rm $out/ex_em_ell_derive/src/lib.rs
+              touch $out/ex_em_ell_derive/src/lib.rs
+            '';
+          };
+        });
 
         ex_em_ell = craneLib.buildPackage (commonArgs // {
           inherit cargoArtifacts;
@@ -67,13 +77,6 @@
 
         packages.ex_em_ell = ex_em_ell;
         packages.default = packages.ex_em_ell;
-
-        # uncomment if there is a binary to be run
-        # apps.cargo-cyclonedx = flake-utils.lib.mkApp {
-        #   drv = packages.ex_em_ell;
-        #   name = "ex_em_ell";
-        # };
-        # apps.default = apps.cargo-cyclonedx;
 
         devShells.default = pkgs.mkShell {
           inputsFrom = builtins.attrValues self.checks.${system};
